GetUnspents use new outputHasRune function
